requirements.txt
asgiref==3.4.1
asttokens==2.0.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.0
dataclasses==0.6
Django==3.0.5
django-cors-headers==3.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
executing==0.8.2
idna==3.3
mysqlclient==2.1.0
pycparser==2.21
Pygments==2.10.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.7
wincertstore==0.2